pyproject.toml
[project]
name = "stargazing"
version = "0.1.0"
description = ""
authors = [
    {name = "zombielu",email = "ludaiche@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11,<3.14"
dependencies = [
    "boto3 (>=1.42.62,<2.0.0)",
    "requests (>=2.32.5,<3.0.0)",
    "apache-airflow (>=3.1.8,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
